#73 Adding Safes model to redux store when loading the app
