+++ src/index.js
@@ -10,11 +10,8 @@ import PageFrame from '~/components/layout/PageFrame'
 import { history, store } from '~/store'
 import theme from '~/theme/mui'
 import AppRoutes from '~/routes'
-import fetchSafes from '~/routes/safe/store/actions/fetchSafes'
 
 import './index.scss'
-
-store.dispatch(fetchSafes())
 
 const Root = () => (
   <Provider store={store}>
