--- src/routes/safe/store/actions/index.js
@@ -1,3 +1,0 @@
-// @flow
-export * from './addSafe'
-export { default as addSafe } from './addSafe'
